+++ src/main/java/com/accounting/app/dto/AttendanceBulkRequest.java
@@ -5,6 +5,8 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import java.util.List;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -12,5 +14,5 @@ public class AttendanceBulkRequest {
     private List<String> employeeIds;
     private Integer month;
     private Integer year;
-    private Double standardDays;
+    private BigDecimal standardDays;
 }
