+++ src/main/java/com/accounting/app/dto/LedgerResponse.java
@@ -3,12 +3,13 @@ package com.accounting.app.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class LedgerResponse {
     private PageResponse<java.util.Map<String, Object>> pageResponse;
-    private double totalDebit;
-    private double totalCredit;
+    private BigDecimal totalDebit;
+    private BigDecimal totalCredit;
 }
